RL/requirements.txt
accelerate
codetiming
datasets
liger-kernel
mathruler
numpy
omegaconf==2.4.0.dev3
pandas
peft
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils
ray[default]
tensordict
torchdata
transformers==4.51.3
vllm==0.8.5
flash-attn==2.6.3
torch==2.6.0
wandb
antlr4-python3-runtime==4.11.1

pyparsing
word2number
timeout_decorator
pebble
sentence_transformers
tensorboard
matplotlib